requirements.txt
streamlit
twilio
google-generativeai
python-dotenv
pandas
scikit-learn
joblib
spacy
flask
